Refactor window show dynamically and change window size
